move no da objects into utils.py
